refactor recheck
